requirements.txt
mcp>=1.20.0
bm25s>=0.2.0
networkx>=3.0
scipy>=1.10.0
pyyaml>=6.0
numpy>=1.24.0
pyjwt>=2.8.0
cryptography>=41.0.0
